pyproject.toml
[project]
name = "browser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.8",
    "browser-use>=0.3.2",
    "uvicorn>=0.34.3",
]
